docs/sample-apps/creator-discovery-app/requirements.txt
pixeltable
twelvelabs
fastapi
uvicorn
python-multipart
pytubefix
Pillow
